Revert moving extensions to separate folder (ProjectExtensions)
