+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/SettingsSection/types.tsx
@@ -1,4 +1,4 @@
-import { ProjectSetupSettings } from '../../../extensions/projectSetup/ProjectSetupSettings'
+import { ProjectSetupSettings } from '../../../projectSetup/ProjectSetupSettings'
 
 export interface ISettingsSectionProps {
   /**
