+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/index.tsx
@@ -5,7 +5,7 @@ import { Pivot, PivotItem } from 'office-ui-fabric-react/lib/Pivot'
 import * as strings from 'ProjectExtensionsStrings'
 import * as React from 'react'
 import { isEmpty } from 'underscore'
-import { ProjectSetupSettings } from '../../extensions/projectSetup/ProjectSetupSettings'
+import { ProjectSetupSettings } from '../../projectSetup/ProjectSetupSettings'
 import { BaseDialog } from '../@BaseDialog'
 import { ExtensionsSection } from './ExtensionsSection'
 import { ListContentSection } from './ListContentSection'
