+++ SharePointFramework/ProjectExtensions/src/projectSetup/deleteCustomizer.ts
