+++ SharePointFramework/ProjectExtensions/src/projectSetup/manifest.json
