+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/CopyListData/index.ts
@@ -1,10 +1,10 @@
 import { stringIsNullOrEmpty, TypedHash } from '@pnp/common'
 import { sp, Web } from '@pnp/sp'
-import { IProjectSetupData } from 'extensions/projectSetup'
+import { IProjectSetupData } from 'projectSetup'
 import { format } from 'office-ui-fabric-react/lib/Utilities'
 import * as strings from 'ProjectExtensionsStrings'
 import { SPField } from 'pp365-shared/lib/models/SPField'
-import { IPlannerTaskSPItem, ListContentConfig, ListContentConfigType } from '../../../../models'
+import { IPlannerTaskSPItem, ListContentConfig, ListContentConfigType } from '../../../models'
 import { BaseTask, BaseTaskError, IBaseTaskParams } from '../@BaseTask'
 import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
 import { ITaskDetails, PlannerConfiguration, TaskAttachment } from '../PlannerConfiguration'
