+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/OnProgressCallbackFunction.ts
