+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/PlannerConfiguration/index.ts
@@ -4,7 +4,7 @@ import { default as MSGraphHelper } from 'msgraph-helper'
 import { format } from 'office-ui-fabric-react/lib/Utilities'
 import { sleep } from 'pp365-shared/lib/util'
 import * as strings from 'ProjectExtensionsStrings'
-import { IProjectSetupData } from 'extensions/projectSetup'
+import { IProjectSetupData } from 'projectSetup'
 import { BaseTask, BaseTaskError, IBaseTaskParams } from '../@BaseTask'
 import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
 import { IPlannerBucket, IPlannerConfiguration, IPlannerPlan } from './types'
