+++ SharePointFramework/ProjectExtensions/src/projectUpgrade/manifest.json
