+++ SharePointFramework/ProjectExtensions/src/templateSelector/index.tsx
@@ -12,8 +12,8 @@ import React from 'react'
 import { render, unmountComponentAtNode } from 'react-dom'
 import HubSiteService from 'sp-hubsite-service'
 import { find, first } from 'underscore'
-import { DocumentTemplateDialog } from '../../components'
-import { SPDataAdapter } from '../../data'
+import { DocumentTemplateDialog } from 'components'
+import { SPDataAdapter } from 'data'
 import { ITemplateSelectorContext, TemplateSelectorContext } from './context'
 import { ITemplateSelectorCommandProperties } from './types'
