+++ SharePointFramework/ProjectExtensions/src/templateSelector/manifest.json
